@@ -38,15 +38,15 @@
     "lint-inspector": "npx @eslint/config-inspector"
   },
   "dependencies": {
-    "@antfu/eslint-config": "5.4.1",
-    "@eslint-react/eslint-plugin": "2.0.6",
-    "@next/eslint-plugin-next": "15.5.4",
-    "@vue/compiler-sfc": "3.5.22",
-    "eslint": "9.37.0",
+    "@antfu/eslint-config": "6.7.1",
+    "@eslint-react/eslint-plugin": "2.3.13",
+    "@next/eslint-plugin-next": "16.0.10",
+    "@vue/compiler-sfc": "3.5.25",
+    "eslint": "9.39.2",
     "eslint-plugin-jsx-a11y": "6.10.2",
     "eslint-plugin-react": "7.37.5",
-    "eslint-plugin-react-hooks": "6.1.1",
-    "eslint-plugin-react-refresh": "0.4.23"
+    "eslint-plugin-react-hooks": "7.0.1",
+    "eslint-plugin-react-refresh": "0.4.26"
   },
   "devDependencies": {
     "@siberiacancode/prettier": "*"
