@@ -1,12 +1,2 @@
-/** @type {import('prettier').Config} */
-module.exports = {
-  printWidth: 100,
-  singleQuote: true,
-  jsxSingleQuote: true,
-  trailingComma: 'none',
-  semi: true,
-  tabWidth: 2,
-  useTabs: false,
-  endOfLine: 'lf',
-  arrowParens: 'always'
-};
+import { prettier } from './index.js';
+export default prettier;
