@@ -1,2 +1,12 @@
-const prettier = require('./.prettierrc.js');
-module.exports = { prettier };
+/** @type {import('prettier').Config} */
+export const prettier = {
+  printWidth: 100,
+  singleQuote: true,
+  jsxSingleQuote: true,
+  trailingComma: 'none',
+  semi: true,
+  tabWidth: 2,
+  useTabs: false,
+  endOfLine: 'lf',
+  arrowParens: 'always'
+};
