@@ -1,2 +1,2 @@
-const { prettier } = require('@siberiacancode/prettier');
-module.exports = prettier;
+import { prettier } from '@siberiacancode/prettier';
+export default prettier;
